+++ chia/cmds/init_funcs.py
@@ -25,7 +25,7 @@ from chia.util.config import (
 from chia.util.ints import uint32
 from chia.util.keychain import Keychain
 from chia.util.path import mkdir
-from chia.util.ssl import (
+from chia.util.ssl_check import (
     DEFAULT_PERMISSIONS_CERT_FILE,
     DEFAULT_PERMISSIONS_KEY_FILE,
     RESTRICT_MASK_CERT_FILE,
